+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 * Delete `showUnit`, `unitStyle`,`contentStyle`, and add `itemBuilder` to DateTimePicker
 * Add `builder` to DateTimePickerUnit
 * Add a new component, `DatePicker`
+* Add a new component, `XSwitch`
 * Remove `horizontalScroll` and `addExpanded`, add `isScrollable`, in `MultiListWheelLinkagePicker` and `MultiListWheelPicker` and related components
 * Change the `PickerEntry` to a `PickerItem`
 * Change the `PickerLinkageEntry` to a `PickerLinkageItem`
 * Removes all Picker `options` initial parameters, which need to be added by yourself
+* Use `ChangedBuilder` instead of `CheckboxState` and `SwitchState`
 
 ## 7.5.2